requirements.txt
Django==3.0.8
requests==2.23.0
shortuuid==1.0.1

django-environ==0.4.5
django-ratelimit==3.0.1
djangorestframework==3.11.0
django-uuslug==1.2.0

Markdown==3.2.2
pymdown-extensions==7.1
mysqlclient==1.4.6
